Update : 총기별 공격력 적용

+++ Assets/Scripts/Interface/IAttackable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IAttackable
 {
-    int GetDamage();
+    float GetDamage();
 }
